remove: Network dependency, unused protocol. move util to WebSocket Module
