+++ AIProject/iCo/Core/Remote/WebSocket/WebSocketClient.swift
@@ -1,49 +1,6 @@
 import Foundation
 import AsyncAlgorithms
 
-public protocol WebSocketProvider {
-    var stateBroadCaster: AsyncStreamBroadcaster<WebSocket.State> { get }
-    var incomingChannel: AsyncChannel<URLSessionWebSocketTask.Message> { get }
-    
-    /// 웹소켓 세션을 연결하고 작업을 생성합니다.
-    func connect() async
-    
-    /// 명시적으로 현재 WebSocket 연결을 정상적으로 종료합니다.
-    ///
-    /// 이 메서드는 서버와의 WebSocket 연결을 `normalClosure` 코드로 닫습니다.
-    func disconnect() async
-    
-    /// 텍스트 형태의 메시지를 WebSocket 서버로 전송합니다.
-    func send(text: String) async throws
-    
-    /// 바이너리(Data) 형태의 메시지를 WebSocket 서버로 전송합니다.
-    func send(data: Data) async throws
-}
-
-public protocol URLSessionType {
-    func makeWebSocketTask(with url: URL) -> WebSocketType
-}
-
-public protocol WebSocketType {
-    var delegate: (any URLSessionTaskDelegate)? { get set }
-    var state: URLSessionTask.State { get }
-    
-    func resume()
-    func cancel(with closeCode: URLSessionWebSocketTask.CloseCode, reason: Data?)
-    func send(_ message: URLSessionWebSocketTask.Message) async throws
-    func cancel()
-    func sendPing(pongReceiveHandler: @escaping @Sendable((any Error)?) -> Void)
-    func receive() async throws -> URLSessionWebSocketTask.Message
-}
-
-extension URLSession: URLSessionType {
-    public func makeWebSocketTask(with url: URL) -> any WebSocketType {
-        return webSocketTask(with: url)
-    }
-}
-
-extension URLSessionWebSocketTask: WebSocketType {}
-
 public class WebSocketClient: NSObject, WebSocketProvider {
     /// 소켓 상태 채널
     private var stateStream: AsyncStream<WebSocket.State>
@@ -98,13 +55,13 @@ public class WebSocketClient: NSObject, WebSocketProvider {
 
     /// 텍스트 형태의 메시지를 WebSocket 서버로 전송합니다.
     public func send(text: String) async throws {
-        guard let task else { throw NetworkError.networkError(URLError(.notConnectedToInternet)) }
+        guard let task else { throw URLError(.notConnectedToInternet) }
         try await task.send(.string(text))
     }
     
     /// 바이너리(Data) 형태의 메시지를 WebSocket 서버로 전송합니다.
     public func send(data: Data) async throws {
-        guard let task else { throw NetworkError.networkError(URLError(.notConnectedToInternet)) }
+        guard let task else { throw URLError(.notConnectedToInternet) }
         try await task.send(.data(data))
     }
 }
@@ -124,7 +81,7 @@ extension WebSocketClient {
                     clearAttempts()
                     receive()
                     checkingAlive()
-                case .failed, .closed:
+                case .closed:
                     debugPrint("Closed")
                     release()
                 case .reconnecting:
@@ -139,7 +96,7 @@ extension WebSocketClient {
     private func receive() {
         receiveTask = Task {
             while true {
-                guard let task else { throw NetworkError.taskCancelled }
+                guard let task else { throw CancellationError() }
                 let message = try await task.receive()
                 await incomingChannel.send(message)
             }
